Extract health check logic to `HealthChecker` type.
